Disable CtrlSpace & update lazy-lock
CtrlSpace doesn't seem to integrate well with some of the other plugins
that are default with LazyVim. For now, in this configuration, I'll give
some of the other options a shot, such as using Telescope for buffer
switching. So far, to get back and forth between files, it seems like
more keys, and typically slower compared to CtrlSpace.