Remove navigator.lua
This enables formatting by default and generally messes with lsp server
settings making it hard to know how servers have been configured. This
also was a reason I spent a huge amount of time trying to figure out why
lua files were not respecting the settings set to the LSP to disable
format on save.